refactor tests into two files
